feat: move incident job to proper module
