+++ app/tests/modules/incident/test_notify_stale_incident_channels.py
@@ -1,10 +1,10 @@
-from jobs import notify_stale_incident_channels
+from modules.incident import notify_stale_incident_channels
 
 from unittest.mock import MagicMock, patch
 
 
 @patch("integrations.slack.channels.get_stale_channels")
-@patch("jobs.notify_stale_incident_channels.log_to_sentinel")
+@patch("modules.incident.notify_stale_incident_channels.log_to_sentinel")
 def test_notify_stale_incident_channels(_log_to_sentinel_mock, get_stale_channels_mock):
     get_stale_channels_mock.return_value = [{"id": "channel_id"}]
     client = MagicMock()
